pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.9",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
]
